+++ lib/rockette.rb
@@ -1,25 +1,26 @@
 # frozen_string_literal: true
 
-require           'json'
-require           'rest-client'
-require           'thor'
-require_relative  'rockette/cli'
-require_relative  'rockette/text'
-require_relative  'rockette/version'
+require           "json"
+require           "rest-client"
+require           "thor"
+require_relative  "rockette/cli"
+require_relative  "rockette/text_helper"
+require_relative  "rockette/version"
 
+# APEX deployment
 module Rockette
   class Error < StandardError; end
 
+  # rest-client calls with error handling and auto retries
   class Rester
-
     VERBS = {
-      'delete' => :Delete,
-      'get'    => :Get,
-      'post'   => :Post,
-      'put'    => :Put
-    }
-  
-    def initialize(headers: {}, meth: 'Get', params: {}, url: 'https://array/', config: {})
+      "delete" => :Delete,
+      "get" => :Get,
+      "post" => :Post,
+      "put" => :Put
+    }.freeze
+
+    def initialize(headers: {}, meth: "Get", params: {}, url: "https://array/", config: {})
       @headers = headers
       @meth    = meth
       @params  = params
@@ -27,55 +28,49 @@ module Rockette
       @config = config
       @config["timeout"] = @config["timeout"] ||= 30
     end
-  
+
     def make_call
-      #@params = @params.to_json unless @meth.downcase == 'get' || @meth.downcase == 'delete'
-      begin
-        response = RestClient::Request.execute(headers: @headers,
-                                               method: VERBS[@meth.downcase],
-                                               payload: @params,
-                                               timeout: @config["timeout"],
-                                               url: @url,
-                                               verify_ssl: false)
-      rescue SocketError => e
-        puts "Socket error, no route to host."
-        puts "#{e.class}: #{e.message}"
-        response = nil
-      rescue => e
-        e.response
-      else
-        response
-      end
+      response = RestClient::Request.execute(headers: @headers,
+                                             method: VERBS[@meth.downcase], payload: @params,
+                                             timeout: @config["timeout"], url: @url, verify_ssl: false)
+    rescue SocketError => e
+      puts "#{e.class}: #{e.message}"
+      nil
+    rescue StandardError => e
+      e.response
+    else
+      response
     end
-  
+
     def cookie
-      if @url =~ /auth\/session/
+      if @url =~ %r{auth/session}
         response = make_call
-        raise 'There was an issue getting a cookie!' unless response.code == 200
-        cookie = (response.cookies.map{|key,val| key + '=' + val})[0]
+        raise "There was an issue getting a cookie!" unless response.code == 200
+
+        (response.cookies.map { |key, val| "#{key}=#{val}" })[0]
       else
-        error_text("cookie", @url.to_s, 'auth/session')
+        error_text("cookie", @url.to_s, "auth/session")
       end
     end
 
     # use rest-client with retry
     def rest_try
-      3.times { |i|
+      3.times do |i|
         response = make_call
-        unless response == nil
+        unless response.nil?
           break response if (200..299).include? response.code
           break response if i > 1
         end
         puts "Failed #{@meth} on #{@url}, retry...#{i + 1}"
         sleep 3 unless i > 1
         return nil if i > 1 # Handles socket errors, etc. where there is no response.
-      }
+      end
     end
 
-  private
+    private
 
     def error_text(method_name, url, wanted)
-      response = {
+      {
         "response" =>
           "ERROR: Wrong url for the #{method_name} method.\n"\
           "Sent: #{url}\n"\
@@ -85,12 +80,10 @@ module Rockette
     end
 
     def responder(response)
-      response = {
+      {
         "response" => JSON.parse(response.body),
         "status" => response.code.to_i
       }
     end
-  
   end
-
 end