+++ lib/rockette/text_helper.rb
@@ -1,10 +1,18 @@
 # frozen_string_literal: true
 
-module Rockette
-
-  def bail_text
+# Add common methods to Rockette module
+module TextHelper
+  def bail
     puts "Bailing, unable to check for application. Can you access the url from here?"
     puts "Double check the url and make sure you have a network connection to the target."
+    exit
   end
 
+  def check_input(input)
+    input unless input.nil?
+  end
+
+  def padder(str)
+    "\n#{str}\n"
+  end
 end