+++ test/rockette_test.rb
@@ -2,6 +2,7 @@
 
 require "test_helper"
 
+# Test Rockette module and classes
 class RocketteTest < Minitest::Test
   def test_that_it_has_a_version_number
     refute_nil ::Rockette::VERSION