+++ CHANGELOG.md
@@ -3,3 +3,9 @@
 ## [0.0.0] - 2021-02-08
 
 - Initial release
+
+
+## [0.0.1] - 2021-03-15
+
+- Unify rest calls into Rester class
+- Improve exception handling including socket errors